internal/relui: re-check blocking issues after release coordinator approval
By now we consistently start a pre-build a week ahead of the planned
release, so we don't get to benefit from having the 'Check blocking
issues' task on the day of the release. Add a second one, that runs
after release coordinator approval, to catch anything that's a release-
blocker by that point. Keep the original one at the start, too, since
it should still be useful to be aware when starting the pre-build.

Also reword a comment added in a previous CL to be more clear.

For golang/go#40279.

Change-Id: Id5c5e337c3e06f2fd38b0161e7dcdc77c301701b
Reviewed-on: https://go-review.googlesource.com/c/build/+/634236
LUCI-TryBot-Result: Go LUCI <golang-scoped@luci-project-accounts.iam.gserviceaccount.com>
Auto-Submit: Dmitri Shuralyov <dmitshur@golang.org>
Reviewed-by: Carlos Amedee <carlos@golang.org>
Reviewed-by: Michael Knyszek <mknyszek@google.com>
Reviewed-by: Dmitri Shuralyov <dmitshur@google.com>