cmd/gorebuild: add basic retries to Get
Gorebuild involves going over the Internet to fetch bytes. Some errors
that it might encounter might be temporary, such that if the caller
were to try again after waiting a bit, it might just succeed. Instead
of trying exactly once before making it something the user has to deal
with, be willing to try a few more times with some delays.

This is intended to reduce the chance of false positive failures in
reports.

Also add a hint for where flags fit to usage text, similarly to how
it's there in the package comment.

For golang/go#57120.

Change-Id: I949fdf49f6317b26326c3c55f93fc6e130fa3839
Reviewed-on: https://go-review.googlesource.com/c/build/+/634655
LUCI-TryBot-Result: Go LUCI <golang-scoped@luci-project-accounts.iam.gserviceaccount.com>
Reviewed-by: Russ Cox <rsc@golang.org>
Reviewed-by: Dmitri Shuralyov <dmitshur@google.com>
Auto-Submit: Dmitri Shuralyov <dmitshur@golang.org>