internal/buildstats: restore implementation for Go 1.23 onwards
CL 584299 was applied to fix build breakage and buy more time. By now,
the breakage has been resolved and so the same implementation can now
be used with any Go version.

The coordinator is still serving the build dashboard and some legacy
builders, and this trivial change is all that's needed for it to be
able to be deployed with Go 1.23 or newer. Apply it now to unblock
that eventual upgrade.

Change-Id: I62f00b5c737aebf3b9f510c15013027f2c68bfb6
Reviewed-on: https://go-review.googlesource.com/c/build/+/636895
LUCI-TryBot-Result: Go LUCI <golang-scoped@luci-project-accounts.iam.gserviceaccount.com>
Auto-Submit: Dmitri Shuralyov <dmitshur@golang.org>
Reviewed-by: David Chase <drchase@google.com>
Reviewed-by: Dmitri Shuralyov <dmitshur@google.com>