cmd/gerritbot/internal/rules: skip checking the wiki repo to minimize friction
Core team reviewers usually seem to approve wiki CLs and PRs without
asking for any changes to a commit message even if it does not conform
to general CL guidelines for the main Go repos.

For example, CL 587895 was approved with the commit title of:
  add "cue" as a well-known struct tag
(without a leading "wiki:" or leading page name or similar).

This seems fine, including I think the general intent was to keep the
reviews very lightweight when the wiki content was moved
(golang/go#61940).

Currently, GerritBot is too picky with wiki PRs.

This CL therefore updates GerritBot to avoid flagging any issues with
wiki PRs.  (An alternative would be to skip just some rules, but it
seems simpler to skip all).

Partly in case there is a later decision to start flagging wiki PRs,
while we are here we update the packageExample and usesTracker functions
to be more specific about the wiki repo (which won't have any effect
right now because the wiki repo is now ignored).

Fixes golang/go#65281

Change-Id: I7f678a9f997f35fa048b78aab1ee4a6d808b4ec6
Reviewed-on: https://go-review.googlesource.com/c/build/+/640135
Reviewed-by: Dmitri Shuralyov <dmitshur@golang.org>
LUCI-TryBot-Result: Go LUCI <golang-scoped@luci-project-accounts.iam.gserviceaccount.com>
Reviewed-by: Carlos Amedee <carlos@golang.org>
Auto-Submit: Dmitri Shuralyov <dmitshur@golang.org>
Reviewed-by: Dmitri Shuralyov <dmitshur@google.com>