cmd/watchflakes/internal/script: lex !~ as a single token
Even though both ~ and !~ operators are documented as supported at
https://go.dev/wiki/Watchflakes, it seems only the former was ever
supported in practice. As it was uncovered when I tried the latter
in https://go.dev/issue/66337#issuecomment-2569872000, it fails to
parse with an "unexpected !" error.

The problem is that lex doesn't find the !~ token, it reports that
sequence as separate tokens ! and ~. Fix its logic.

Fixes golang/go#71119.

Change-Id: I08dd845a59e976a5eb2687924dce972680c90077
Reviewed-on: https://go-review.googlesource.com/c/build/+/640275
Auto-Submit: Dmitri Shuralyov <dmitshur@golang.org>
Reviewed-by: Dmitri Shuralyov <dmitshur@google.com>
Reviewed-by: Cherry Mui <cherryyz@google.com>
LUCI-TryBot-Result: Go LUCI <golang-scoped@luci-project-accounts.iam.gserviceaccount.com>