cmd/updatestd: skip repos without internal-branch.go1.n-vendor branch
In our previous process, we used to preemptively create an
internal-branch.go1.n-vendor branch for every vendored golang.org/x
repo that's a direct dependency of either the std or cmd module.
But now we create those branches on demand, only when they're needed.
So, stop treating a missing branch as an error to report, even if it
is a direct dependency of std or cmd.

For golang/go#36905.

Change-Id: Id5d2b3b34b286ae7d01b1b9e48c6d98d7c72df07
Reviewed-on: https://go-review.googlesource.com/c/build/+/643517
Reviewed-by: Dmitri Shuralyov <dmitshur@google.com>
Auto-Submit: Dmitri Shuralyov <dmitshur@golang.org>
Reviewed-by: Cherry Mui <cherryyz@google.com>
LUCI-TryBot-Result: Go LUCI <golang-scoped@luci-project-accounts.iam.gserviceaccount.com>