main.star: update supported Go versions and drop release-branch.go1.22
The Go 1.24.0 release is out, so 1.23 is the oldest supported release.¹
The previous infrastructure (cmd/coordinator) does this automatically,
whereas here I authored this CL manually. We'll explore automating it
later on², possibly with relui generating the CL.

CL 634335 added NEXT_GO when the release-branch.go1.24 branch was cut
along with the release of the first Go 1.24 release candidate. It was
documented that it's to be deleted later. But we'll need to re-add it
or something equivalent when release-branch.go1.25 will be created so
just document it and keep it around, even if it's unused at this time.
Rename it to 'next-up' for better vertical alignment.

¹ https://go.dev/doc/devel/release#policy
² I said the same thing 6 months ago in CL 605295. That said, there
  have been various minor refactors here, including creating tracking
  constants in CL 611941 and CL 634335, so any automated workflow would
  have needed to be manually updated too. We can still eventually do it
  once these every-6-month update CLs become overly repetitive.

Change-Id: I53e399daad44d443c4a50070ed5f44d34b14037f
Reviewed-on: https://go-review.googlesource.com/c/build/+/648675
LUCI-TryBot-Result: Go LUCI <golang-scoped@luci-project-accounts.iam.gserviceaccount.com>
Reviewed-by: Dmitri Shuralyov <dmitshur@google.com>
Reviewed-by: Cherry Mui <cherryyz@google.com>
Auto-Submit: Dmitri Shuralyov <dmitshur@golang.org>