internal/task: add GenerateAutoSubmitChange method to CloudBuildClient
I originally planned to use the existing CloudBuildClient.RunScript API
for this task, similarly to how it's used for monthly x/ repo tagging.
One key limitation of the RunScript API is that it requires the caller
to pass the exact files that will be copied out to Google Cloud Storage.
That works fine if go.mod and go.sum files were the only ones to change,
but 'go fix ./...' may update arbitrary other files too.

It's not too hard to expand RunScript to be able to copy directories
recursively, but that means it'd be copying the entire x/ repo content
even if only a few files changed. It's possible to try to detect which
files changed, and copy only those, or to create a .patch file.

However, we're already relying on the Cloud Build to clone the repo and
run commands to generate files, so it can work well for the Cloud Build
invocation to also take on the responsibility of mailing the auto-submit
CL, avoiding the need to copy anything to temporary storage only so that
relui can do the mailing itself. Also reuse git-generate as a building
block, pinning a known version for stability.

The result is an API that's very easy to use, and has the benefit of
including the script that describes the change in the commit message.

For golang/go#69095.

Change-Id: I78bb69567044e334fea693bf32a02de86696e7c9
Reviewed-on: https://go-review.googlesource.com/c/build/+/651336
Auto-Submit: Dmitri Shuralyov <dmitshur@golang.org>
Reviewed-by: Dmitri Shuralyov <dmitshur@google.com>
Reviewed-by: Cherry Mui <cherryyz@google.com>
LUCI-TryBot-Result: Go LUCI <golang-scoped@luci-project-accounts.iam.gserviceaccount.com>