internal/task: add GoDirectiveXReposTasks to maintain go directive
In general, the structure of GoDirectiveXReposTasks is similar to that
of TagXReposTasks, which performs a related task. It's simpler because
the go directive can be upgraded in all golang.org/x repos in parallel,
so the complexity of traversing the module graph as part of finding an
optimal order doesn't come up.

The SelectRepos method uses the canonical source of golang.org/x repos
(the x/build/repos package) and includes all of them, except ones that
opt out from this behavior via the AutoMaintainGoDirective field.

The BuildPlan method schedules tasks to send CLs for each x/ repo, and
then awaits their submission.

MaintainGoDirectiveAndMailCL has the logic to construct a git-generate
script for the given x/ repository with go directives to be maintained,
and to create and mail the generated CL using GenerateAutoSubmitChange.

Add coverage for the new behavior to TestRelease.

Fixes golang/go#69095.

Change-Id: I8729539a690e37d81f02b49625251ae86e01e79a
Reviewed-on: https://go-review.googlesource.com/c/build/+/649676
Auto-Submit: Dmitri Shuralyov <dmitshur@golang.org>
Reviewed-by: Dmitri Shuralyov <dmitshur@google.com>
Reviewed-by: Cherry Mui <cherryyz@google.com>
LUCI-TryBot-Result: Go LUCI <golang-scoped@luci-project-accounts.iam.gserviceaccount.com>