internal/task: narrow mailing list email search by subject and keywords
The general approach relui takes for handling (pre-)announcements sent
to a mailing list such as golang-announce is as follows: it sends the
email and waits for it to appear there, finding out its URL. For this
to work, it relies on the simple invariant that each announcement email
we send has a subject that uniquely identifies said email.

For example, if the subject is "[security] Go 1.18.3 and Go 1.17.11 are
released", that's enough information for it to be possible to find out
whether that announcement email is already sent, and what its URL is.
It also lets us have a safety check to avoid an email with duplicate
content from being sent multiple times by accident.

This worked well for various release announcements because they tend to
know the version to be released, and that version being in the subject
guaranteed uniqueness. This doesn't scale¹ for "[security] {{.Module}}
fix pre-announcement" pre-announcements.

To deal with this, it should work to rely on including unique keywords
in the search for sent email, at least for ones where the subject alone
isn't unique enough. We'll know if this ever breaks or needs maintenance
because this makes it viable to make TestFindGoogleGroupsThread into a
test that always runs, rather than only when a human runs it, inspecting
its output manually.

This fixes the current TestPreAnnounceXFix failure where it's matching
a recent real email unintentionally, and makes it viable for x/ repo
pre-announcements to continue to await the email it sent and find out
its the URL.

¹ Maybe in the future we'll get to a point where the automation can know
  in advance what the module version and put that in the subject. That's
  tricky because at minimum it needs to take into account the monthly x/
  repo tagging automation, which is currently a separate relui workflow.

Change-Id: Ic0602225d59e584e10715b00f22e55baf17ff297
Reviewed-on: https://go-review.googlesource.com/c/build/+/651335
LUCI-TryBot-Result: Go LUCI <golang-scoped@luci-project-accounts.iam.gserviceaccount.com>
Auto-Submit: Dmitri Shuralyov <dmitshur@golang.org>
Reviewed-by: Dmitri Shuralyov <dmitshur@google.com>
Reviewed-by: Carlos Amedee <carlos@golang.org>