internal/task: detect "no CL needed" in GenerateAutoSubmitChange
The API of GenerateAutoSubmitChange aspirationally promised that:

// If the requested contents match the state of the repository,
// no change is created and the returned change ID will be empty.

That edge case typically doesn't come up during the task of maintaining
the go directive, so it was hard to test. Now that the happy path where
there are generated files works well, it's easier to add handling for
the "nothing changed" case. In particular, the UpdateBundle task of
BundleNSSRootsTask (which is migrated to the new API in the next CL)
makes for a convenient way to test this out both for the real Cloud
Build client and the fake.

For golang/go#69095.

Change-Id: I9a4271f2dd7b463433403190d5588d39b8416751
Reviewed-on: https://go-review.googlesource.com/c/build/+/651736
Auto-Submit: Dmitri Shuralyov <dmitshur@golang.org>
LUCI-TryBot-Result: Go LUCI <golang-scoped@luci-project-accounts.iam.gserviceaccount.com>
Reviewed-by: Dmitri Shuralyov <dmitshur@google.com>
Reviewed-by: Cherry Mui <cherryyz@google.com>