internal/task: use relui service account instead of secret for git authn
Remove reliance on the secret in Secret Manager in one more place. This
is easier to do now that the GenerateAutoSubmitChange implementation is
done and handles all of the intended edge cases; it was very helpful be
able to iterate with a shorter feedback loop during earlier development
stages.

For golang/go#69095.

Change-Id: I103278cfde8df10c8a9f0807607a2a545fa2e376
Reviewed-on: https://go-review.googlesource.com/c/build/+/658018
Reviewed-by: Roland Shoemaker <roland@golang.org>
Reviewed-by: Dmitri Shuralyov <dmitshur@google.com>
LUCI-TryBot-Result: Go LUCI <golang-scoped@luci-project-accounts.iam.gserviceaccount.com>
Auto-Submit: Dmitri Shuralyov <dmitshur@golang.org>