perf: fix homepage chart header links
On the homepage, where no branch is specified in the query parameter,
the generated link for each chart's header will specify branch=null as a
query parameter. This is of course broken, and leads to a failure to
fetch git history when clicking the link.

Let's assume the same thing the server does when no branch is specified,
and that's that the branch is "master."

Thanks to Carlos Amedee for noticing this and letting me know.

Change-Id: I0bb0675dcce7f5129c908cf7ac213a5e3387a242
Reviewed-on: https://go-review.googlesource.com/c/build/+/656015
Reviewed-by: Carlos Amedee <carlos@golang.org>
LUCI-TryBot-Result: Go LUCI <golang-scoped@luci-project-accounts.iam.gserviceaccount.com>