internal/releasetargets: update list for Go 1.25
The release freeze for Go 1.25 is a bit over a month away.
Do a preliminary update of the release target list.
So far it is the same as for Go 1.24, with the following diff:

- The new minimum macOS version is macOS 12.

We'll revisit this after the freeze in case anything changes.

For golang/go#40561.
For golang/go#69839.

Change-Id: I26bdcaa88b39b6d6342dbb881e53f3035862a24f
Reviewed-on: https://go-review.googlesource.com/c/build/+/664215
LUCI-TryBot-Result: Go LUCI <golang-scoped@luci-project-accounts.iam.gserviceaccount.com>
Reviewed-by: Dmitri Shuralyov <dmitshur@google.com>
Reviewed-by: Carlos Amedee <carlos@golang.org>
Auto-Submit: Dmitri Shuralyov <dmitshur@golang.org>