cmd/gomote: fix racy group updates from concurrent gomote processes
Currently the updates to the group JSON files are racy, and can result
in all sorts of problems when multiple gomote processes try to update
the same group metadata.

Fix this by using a directory structure to represent groups, and rely
on filesystem atomicity to make concurrent updates safe. A group is now
represented as a directory, <group>.instances, in the config directory.
Each directory contains an empty file per instance, with the file's name
being the instance name.

This has one problem where if instances are not explicitly destroyed but
instead time out, the group directory structure may contain stale
entries. This is especially problematic because instance names are
non-unique and are frequently reused. (mknyszek-gotip-linux-amd64-0 will
be reused if there are no instances under my name.) To deal with this,
when we create a new instance, we always prune the instance name from
all other groups before adding it to any group for the first time.

Change-Id: I06a88302985e810c6e676dd8dd7838c55e79e46c
Reviewed-on: https://go-review.googlesource.com/c/build/+/664899
Auto-Submit: Michael Knyszek <mknyszek@google.com>
LUCI-TryBot-Result: Go LUCI <golang-scoped@luci-project-accounts.iam.gserviceaccount.com>
Reviewed-by: Carlos Amedee <carlos@golang.org>