main.star: support blocking x/ repo SlowBots in the main Go repo
It's possible for a CL that lands in the main Go repo to contain a bug
that is caught in post-submit by tests in one of the golang.org/x repos.
(See a recent example at go.dev/issue/73317.)

It's convenient to be able to confirm in pre-submit that a fix CL will
indeed fix the failing x/ repo tests. It's already possible to trigger
an optional SlowBot (see https://ci.chromium.org/b/8717939122895000929
for example), but it turned out not to be possible to make said SlowBot
blocking via the Cq-Include-Trybots trailer:

	Failed to compute tryjob requirement. Reason:

	builder "golang/try/x_review-gotip-windows-amd64-longtest"
	is included but not defined in the LUCI project

There's no reason for us to disallow that. Add support for it.

Change-Id: I571f9774f4a26a0ef93b9b182fc3a1ee85d2f4e9
Reviewed-on: https://go-review.googlesource.com/c/build/+/665017
Reviewed-by: Dmitri Shuralyov <dmitshur@google.com>
LUCI-TryBot-Result: Go LUCI <golang-scoped@luci-project-accounts.iam.gserviceaccount.com>
Reviewed-by: Michael Knyszek <mknyszek@google.com>
Auto-Submit: Dmitri Shuralyov <dmitshur@golang.org>