cmd/bootstrapswarm: allow hostname flag to be left empty on GCE
CL 526618 changed the 'on GCE' path of bootstrapswarm to forcibly use
its own internal hostname computation rather than the value of the
hostname flag, dropping "(required)" from its description. But the flag
is still required when not on GCE, and even on GCE bootstrapswarm would
exit fatally if os.Getenv("HOSTNAME") happened to be the empty string.

Make all of this more precise. Spotted while investigating why the
swarming bot wouldn't come up in sid and wanted to eliminate this
possibility, even though it turned out to be due to a Python 3.13
incompatibility as described in CL 665016 and not this.

Change-Id: Ibf65057290e9d8cb79a2800fdc982d35d7edf68b
Reviewed-on: https://go-review.googlesource.com/c/build/+/666375
LUCI-TryBot-Result: Go LUCI <golang-scoped@luci-project-accounts.iam.gserviceaccount.com>
Auto-Submit: Dmitri Shuralyov <dmitshur@golang.org>
Reviewed-by: Dmitri Shuralyov <dmitshur@google.com>
Reviewed-by: Michael Pratt <mpratt@google.com>