internal/task: consider failing to set a review a soft failure, part 1/2
For the purpose of CreateAutoSubmitChange, the most important outcome
is whether a CL is created. It's possible that adding reviewers, a step
that happens as part of setting the review votes, may fail if there's
a change in Gerrit accounts that hasn't propagated to the gophers
package, something we've run into a couple of times.

Since it may always happen in the future, consider it a soft failure,
opting to let the release coordinators handle adding reviewers and/or
setting review votes. This avoids the friction of having to redeploy
relui, the alternative way to handle such cases.

While here, also use the new ID field of the created change instead of
re-creating it ourselves.

Fixes golang/go#68276.

Change-Id: I036cfeaeef9be44c4d8afd12c46aed2d71342a1d
Reviewed-on: https://go-review.googlesource.com/c/build/+/670475
Reviewed-by: Dmitri Shuralyov <dmitshur@google.com>
Reviewed-by: Carlos Amedee <carlos@golang.org>
LUCI-TryBot-Result: Go LUCI <golang-scoped@luci-project-accounts.iam.gserviceaccount.com>
Auto-Submit: Dmitri Shuralyov <dmitshur@golang.org>