internal/task: include detailed Cloud Build log URL upon failure
In the unexpected case that a cloud build script fails, it's generally
useful to see the log that contains the commands and their output, and
not the result GCS bucket. That bucket likely holds no results anyway.

The error before was:

	build "daacc1b0-0802-46d6-b564-0d2dfe29397b" failed,
	see gs://golang-release-staging/relui-scratch/build-outputs/script-build-8224756220732546540: …

And after:

	build "daacc1b0-0802-46d6-b564-0d2dfe29397b" failed,
	see https://console.cloud.google.com/cloud-build/builds/daacc1b0-0802-46d6-b564-0d2dfe29397b?project=872405196845: …

I found this while working on GenerateAutoSubmitChange in CL 651336,
and it's already applied there. Apply this improvement to RunScript too.

Change-Id: Id9ecaf84d0e49bd89cd1078a6dde57ab67064ef3
Reviewed-on: https://go-review.googlesource.com/c/build/+/671055
Reviewed-by: Dmitri Shuralyov <dmitshur@google.com>
LUCI-TryBot-Result: Go LUCI <golang-scoped@luci-project-accounts.iam.gserviceaccount.com>
Reviewed-by: Hongxiang Jiang <hxjiang@golang.org>
Auto-Submit: Dmitri Shuralyov <dmitshur@golang.org>