main.star: add linux-s390x-race builder
The previous build infrastructure had a -race builder for linux/s390x,
so port it over too.

Notably, the previous configuration enabled the -race configuration on
the main branch of the main Go repo only:

	Name:      "linux-s390x-ibm-race",
	buildsRepo: func(repo, branch, goBranch string) bool {
		return repo == "go" && goBranch == "master"
	},

It is possible to implement that exact rule here too, but it would add
special configuration to maintain. Try the default behavior first, and
we can revisit this if it proves to be needed.

For golang/go#67307.

Change-Id: I09012ae7f7d648d22da62e6986228ed856720902
Reviewed-on: https://go-review.googlesource.com/c/build/+/672075
LUCI-TryBot-Result: Go LUCI <golang-scoped@luci-project-accounts.iam.gserviceaccount.com>
Reviewed-by: Michael Knyszek <mknyszek@google.com>
Reviewed-by: Dmitri Shuralyov <dmitshur@google.com>
Auto-Submit: Dmitri Shuralyov <dmitshur@golang.org>