internal/migration: stop running more builders
We've considered porting these builders but decided not to at this time.
We have a standard Debian builder in many run-mod variations, and a Sid
builder which aims to use a newer Linux kernel version. If we encounter
compelling motivation to have more per-distro builders, we will revisit
this. Stop the legacy builders now.

Also mark windows-arm-zx2c4 as something that's not going to be ported,
as the windows/arm port is slated for removal.

For golang/go#60468.
For golang/go#67761.

Change-Id: I595ff2df382130f6eaeca8f1a163f4939fc4873e
Reviewed-on: https://go-review.googlesource.com/c/build/+/672216
Reviewed-by: Cherry Mui <cherryyz@google.com>
LUCI-TryBot-Result: Go LUCI <golang-scoped@luci-project-accounts.iam.gserviceaccount.com>
Auto-Submit: Dmitri Shuralyov <dmitshur@golang.org>
Reviewed-by: Dmitri Shuralyov <dmitshur@google.com>