cmd/racebuild: simplify and fix Windows build
Currently the Windows build script for the race detector syso involves
an external package manager to install git, and does some fiddling with
the environment to clear everything and add back in a C compiler. This
all appears to be unnecessary. Both a C compiler and git are in PATH
already on LUCI.

For golang/go#72767.

Change-Id: I3e9ee85367ce08b06d80eabc265e1d1839815c5a
Reviewed-on: https://go-review.googlesource.com/c/build/+/676500
Reviewed-by: Dmitri Shuralyov <dmitshur@google.com>
Auto-Submit: Michael Knyszek <mknyszek@google.com>
LUCI-TryBot-Result: Go LUCI <golang-scoped@luci-project-accounts.iam.gserviceaccount.com>
Reviewed-by: Dmitri Shuralyov <dmitshur@golang.org>