internal/installer/darwinpkg: speed up find command significantly
I noticed that installing Go on my personal MacBook Pro M4 (macOS 15.5)
was very slow in the “Running package scripts” step.

I was curious and discovered the find command I’m changing.
Running this find command is very slow on my mac for some reason:

  sudo time bash -c 'cd /usr/local/go && find . -exec echo \{\} \;'

I aborted the command after over a minute (!!).

I was wondering if the command would be faster when reducing
subprocess execution overhead (by using the + syntax),
and indeed that reduces the total time to a fraction of a second!

But how widely available is the + syntax on Mac?
Looking at the blame layer of the find(1) man page:
https://github.com/apple-oss-distributions/shell_cmds/blame/b33f386ac431f17d4799b5662f20ff2ddf0773b9/find/find.1#L4
…we can see that the + syntax was introduced with shell_cmds-116
(don’t get distracted by the git commit date, it looks incorrect),
which apparently was introduced with Mac OS X 10.5 (in 2007):
https://git.saurik.com/apple/shell_cmds.git/history/26812dd2205b1e490a62d45c74d0b14e5eb24ad7:/uname
Given that Go requires macOS 11+, we can make use of this syntax.

Change-Id: I1acc051e47cbedbe3d6fd9b44f46f4193d1cdd13
Reviewed-on: https://go-review.googlesource.com/c/build/+/681815
Reviewed-by: Dmitri Shuralyov <dmitshur@google.com>
Reviewed-by: Dmitri Shuralyov <dmitshur@golang.org>
LUCI-TryBot-Result: Go LUCI <golang-scoped@luci-project-accounts.iam.gserviceaccount.com>