internal/installer/darwinpkg: detect amd64 installer on arm64 host
Both of our installers are correctly annotated for the intended host
architecture. That is, the hostArchitectures option is set to x86_64
on .darwin-amd64.pkg installers, and to arm64 on .darwin-arm64.pkg ones.

When an Apple silicon device encounters a .darwin-amd64.pkg installer,
it doesn't have a way of knowing that Go also offers another dedicated
installer, and that it's most likely a user mistake of using the wrong
Go installer. In a pursuit of backwards compatibility, it tries to run
the installer via Rosetta 2, and users may not notice their mistake.

There may be a valid reason to want to use or test the darwin/amd64 port
on Apple silicon devices, but that's a very niche and advanced use case.
In practice, it's more likely to happen unintentionally. Don't try to
support it - advanced users who seek this rare behavior can work around
it by building from source or using the binary archive. Instead, detect
it using the mechanism documented at
https://developer.apple.com/documentation/apple-silicon/about-the-rosetta-translation-environment#Determine-Whether-Your-App-Is-Running-as-a-Translated-Binary
and report it as an error.

Fixes golang/go#59010.

Change-Id: I737d1634f82671a4523af9c7a99e8c6497ce844c
Reviewed-on: https://go-review.googlesource.com/c/build/+/682015
LUCI-TryBot-Result: Go LUCI <golang-scoped@luci-project-accounts.iam.gserviceaccount.com>
Auto-Submit: Dmitri Shuralyov <dmitshur@golang.org>
Reviewed-by: Dmitri Shuralyov <dmitshur@google.com>
Reviewed-by: Michael Stapelberg <stapelberg@google.com>