internal/gomote: add swarming task link to gomoteserver UI
The gomoteserver status page lists active instances but does not
provide a direct link to the underlying swarming task for instances
created via swarming. This makes it difficult to find and debug the
associated task in the Swarming UI. This change adds a link to the
corresponding swarming task used to create each gomote instance. A
link to query for all gomote related swarming tasks has been added to
the footer of the main gomote server page.

Change-Id: Ief2f8ff78a9e61646effe191a6d49768ce9b6b66
Reviewed-on: https://go-review.googlesource.com/c/build/+/687235
Reviewed-by: Michael Knyszek <mknyszek@google.com>
LUCI-TryBot-Result: Go LUCI <golang-scoped@luci-project-accounts.iam.gserviceaccount.com>