cmd/gomote: add executable bits for .bash and .rc files on Windows
On Windows, the FileInfo for a file will not have executable bits, so if
we read the file info for make.bash, and then set that mode bit in the
tar file we send to the mote, we'll lose any executable bits that are in
the git repo. This results in make.bash and related files being written
without executable bits on the mote. Add u+x to the mode on .bash and
.rc files on Windows, all of which files in GOROOT are executable in git
as of when this CL was created.

Change-Id: I558668b8cbaade07538c00aee3253b2cdb969a73
Reviewed-on: https://go-review.googlesource.com/c/build/+/688535
Reviewed-by: Cherry Mui <cherryyz@google.com>
Reviewed-by: Dmitri Shuralyov <dmitshur@golang.org>
Auto-Submit: Michael Matloob <matloob@golang.org>
Reviewed-by: Carlos Amedee <carlos@golang.org>
LUCI-TryBot-Result: Go LUCI <golang-scoped@luci-project-accounts.iam.gserviceaccount.com>
Reviewed-by: Dmitri Shuralyov <dmitshur@google.com>