cmd/gomote: use the major release version in creating dummy version
One of the sweet benchmarks (go-build) checks agains the go version it's
benchmarking to see if it's new enough to run the benchmark. Create
versions that can be compared and that use the major release version
from goversion.go, to allow that benchmark to consider a dummy gomote
version.

Change-Id: Id76db5833f7165c4f99e5b2a32075262978b490b
Reviewed-on: https://go-review.googlesource.com/c/build/+/689635
Reviewed-by: Dmitri Shuralyov <dmitshur@golang.org>
Reviewed-by: Dmitri Shuralyov <dmitshur@google.com>
Reviewed-by: Michael Knyszek <mknyszek@google.com>
LUCI-TryBot-Result: Go LUCI <golang-scoped@luci-project-accounts.iam.gserviceaccount.com>