internal/task: update announce-rc test case to use go1.25rc1
The announce-rc.md template has a section that was used for go1.23rc1.
It has no effect on future RC announcements, so I'll delete it in the
next CL. Update its test case to target a newer RC version in this CL
to make the diff of the next CL easier to review.

For golang/go#71574.

Change-Id: Ice713ac7865b164e29e50c69b6becdb164a814b7
Reviewed-on: https://go-review.googlesource.com/c/build/+/692235
Reviewed-by: Dmitri Shuralyov <dmitshur@google.com>
Reviewed-by: Robert Findley <rfindley@google.com>
Auto-Submit: Dmitri Shuralyov <dmitshur@golang.org>
LUCI-TryBot-Result: Go LUCI <golang-scoped@luci-project-accounts.iam.gserviceaccount.com>