internal/task: add suggestion to opt in to Go telemetry in RCs
We've experimented with including a suggestion to opt in to Go telemetry
in the Go 1.23 RC 1 release announcement. That was a one-off inclusion,
but it has worked well. In CL 692235, Rob has suggested we try including
it in future RC announcements too.

Bring it back. Users who are already helping test the release candidates
ahead of time are more likely to be interested in helping by opting in,
and this will serve as a regular reminder. We can rely on the telemetry
subcommand being present in all supported Go releases by now.

Change-Id: I3200725ac0c35350b8a6c0aa1961fd1a8884b3d1
Reviewed-on: https://go-review.googlesource.com/c/build/+/692375
Reviewed-by: Dmitri Shuralyov <dmitshur@google.com>
LUCI-TryBot-Result: Go LUCI <golang-scoped@luci-project-accounts.iam.gserviceaccount.com>
Reviewed-by: Robert Findley <rfindley@google.com>
Auto-Submit: Dmitri Shuralyov <dmitshur@golang.org>