internal/workflow: convert unexpected panics in tasks to task errors
Relui tries to be as resilient to failure as possible, enabling the
release coordinators to retry individual tasks that happen to need
multiple tries before they succeed, and move forward.

Tasks are the smallest unit of work that can be retried. Most tasks
are automatically tried up to 3 times, and even then, a human release
coordinator may inspect the situation and manually retry again if it's
deemed safe. (The alternative paths might be that the workflow is
started anew with different inputs, or task code is modified before
retrying, and so on.)

When a task encounters a normal failure mode, it reports that with an
error. Although rare, it is also possible for a task to run into a
runtime panic: possibly due to a bug in the code, or in one of the
dependencies being used.

Currently such panics are handled at the the process level: whenever
one of the tasks panics, the relui process is restarted, and each task
that was running at the time is placed in a "task interrupted before
completion" state with the option to manually retry to keep going.

This change adds panic handling to the workflow package, so that if
one particular task happens to run into a rare panic, that failure
will be contained to said task without also taking down all other
unrelated tasks that merely happen to be running at the same time.

For golang/go#58228.
For golang/go#58456.

Change-Id: Iafd72cdf1ebb3b3f1fe7109cee836d0a741ecfb9
Reviewed-on: https://go-review.googlesource.com/c/build/+/692655
LUCI-TryBot-Result: Go LUCI <golang-scoped@luci-project-accounts.iam.gserviceaccount.com>
Reviewed-by: Dmitri Shuralyov <dmitshur@google.com>
Reviewed-by: Mark Freeman <markfreeman@google.com>
Auto-Submit: Dmitri Shuralyov <dmitshur@golang.org>