internal/relui: update TestSecurity target to go1.24.1 from go1.18rc1
Change TestSecurity to make a minor release of a supported Go version.
The next CL adds an internet-requiring task to the RC release kind,
so it would necessitate adding a skip to this test in short mode. But
other than that, it's fairly fast (5 seconds or so), and minor releases
also work as a representative release kind with private-track security
fixes, so switch to that to let it keep running in short mode.

CL 664216 did a similar modernization update for TestRelease previously.

Change-Id: I460e76ebbcf39a8d5100021108ea50fa2ac682e2
Reviewed-on: https://go-review.googlesource.com/c/build/+/695215
LUCI-TryBot-Result: Go LUCI <golang-scoped@luci-project-accounts.iam.gserviceaccount.com>
Reviewed-by: Roland Shoemaker <roland@golang.org>
Reviewed-by: Dmitri Shuralyov <dmitshur@google.com>
Auto-Submit: Dmitri Shuralyov <dmitshur@golang.org>