main.star: compose presubmit filters via intersection
Currently presubmit filters are compose via a union-like operation,
which only ever expands the set of situations that a builder is run in.
But the filters should instead compose to be more conservative. For
example, this means today that longtest-race builders run in many
situations where they're very much overkill, and they take a *very* long
time to run.

That's the immediate target of this change: to run longtest-race
builders less often in presubmit. But an intersection based policy for
filters is in general more intuitive. Note however that the intersection
policy here relies on an exact match on filters, so it's fairly
rudimentary and likely overly conservative in the grand scheme of
things, but the composition of longtest and race is essentially the only
place this matters today, so it's OK.

Change-Id: Ie823fc1f8cb593e8b468d24fc89a488090d0e9d8
Reviewed-on: https://go-review.googlesource.com/c/build/+/700597
LUCI-TryBot-Result: Go LUCI <golang-scoped@luci-project-accounts.iam.gserviceaccount.com>
Reviewed-by: Dmitri Shuralyov <dmitshur@golang.org>
Reviewed-by: Dmitri Shuralyov <dmitshur@google.com>
Auto-Submit: Michael Knyszek <mknyszek@google.com>