main.star: make -longtest-race builders optional in presubmit
CL 700597 removed presubmit filters for this runmod combination, and we
also want to also take additional action to override their individual
"presubmit on" from being used as is, since the combination of these
is too expensive to be included in presubmit by default for all CLs on
the main branch. Any one of them in isolation should keep running as
they did before, of course.

Do this for this particular pair directly now. If/when we want to move
this to be more general, one idea is each runmod could have a heavy bit
(or a weight scale), and we disable presubmit when their sum or product
is beyond some threshold (i.e., no more than one very expensive runmod
at a time) - something along the lines of CL 700898.

Change-Id: Ib30f075b436ff6b264380e89783b7c3c914fecc5
Reviewed-on: https://go-review.googlesource.com/c/build/+/700897
Reviewed-by: Michael Knyszek <mknyszek@google.com>
Auto-Submit: Dmitri Shuralyov <dmitshur@golang.org>
LUCI-TryBot-Result: Go LUCI <golang-scoped@luci-project-accounts.iam.gserviceaccount.com>
Reviewed-by: Dmitri Shuralyov <dmitshur@google.com>