internal/releasetargets: update list for Go 1.26
The release freeze for Go 1.26 is a bit over a month away.
Do a preliminary update of the release target list.
So far it is the same as for Go 1.25, with the following diff:

- The freebsd/riscv64 port is not included because it's marked broken.

We'll revisit this after the freeze in case anything changes.

For golang/go#40561.
For golang/go#73568.

Change-Id: If95d3cad31b6d6c39b03058e5066ece8cebbf661
Reviewed-on: https://go-review.googlesource.com/c/build/+/713560
Reviewed-by: David Chase <drchase@google.com>
LUCI-TryBot-Result: Go LUCI <golang-scoped@luci-project-accounts.iam.gserviceaccount.com>
Auto-Submit: Dmitri Shuralyov <dmitshur@golang.org>
Reviewed-by: Dmitri Shuralyov <dmitshur@google.com>